Implemented a useless GAPI
Temporarily added code that called YouTube API to check for a video's licensedContent field, assuming that this field indicated if the video contained copyrighted music or not. It does not mean that. It is some other bizarre thing which isn't useful for determining whether or not a video contains DMCA. The scraper implementation wins out here.