Added link to tutorial, fixed empty span problem
Added a link to the tutorial video on my YouTube, and added some css that fixed the youtube video input span not being editable when cleared and losing focus. however, this required adding display: inline-block, which makes the page not fit in one screen anymore, so I slightly reduced the font size of links to make it all fit again.